chore: move helper functions of `utils.py` into `setup.py`
